Fixed missing brightmaps when using texture quality reduction. This was caused in the creation of FMaterial

+++ src/common/rendering/vulkan/textures/vk_hwtexture.cpp
@@ -207,7 +207,7 @@ void VkHardwareTexture::CreateImage(FTexture *tex, int translation, int flags)
 					dataPos += mipSize;
 					mipWidth = std::max(1u, (mipWidth >> 1));
 					mipHeight = std::max(1u, (mipHeight >> 1));
-					mipSize = std::max(1u, ((mipWidth + 3) / 4)) * std::max(1u, ((mipHeight + 3) / 4)) * 16;
+					mipSize = (size_t)std::max(1u, ((mipWidth + 3) / 4)) * std::max(1u, ((mipHeight + 3) / 4)) * 16;
 				}
 			}
 			else {
